Utilizado validação nas rotas necessárias

+++ back-end/src/controllers/CarsController.ts
@@ -1,12 +1,15 @@
 import { Request, Response, NextFunction } from 'express';
 import CarsService from '../services/CarsService';
-import { CREATED_STATUS, OK_STATUS } from '../helpers/HTTPCodes';
+import { CREATED_STATUS, OK_STATUS, UNAUTHORIZED_STATUS } from '../helpers/HTTPCodes';
 import { BaseCar } from '../helpers/interfaces';
+import { USER_UNAUTHORIZED_ERROR } from '../helpers/errorsCodes';
 
 class CarsController {
   async create(
     req: Request, res: Response,
   ): Promise<Response | void> {
+    const { userRole } = req.user;
+    if (userRole !== 'admin') return res.status(UNAUTHORIZED_STATUS).json(USER_UNAUTHORIZED_ERROR);
     const { carModel, costHour }: BaseCar = req.body;
     const car = await CarsService.create(
       { carModel, costHour },
@@ -34,6 +37,10 @@ class CarsController {
     req: Request, res: Response, next: NextFunction,
   ): Promise<Response | void> {
     const { id } = req.params;
+    const { userRole, userId } = req.user;
+    if (userRole !== 'admin' && id !== userId) {
+      return res.status(UNAUTHORIZED_STATUS).json(USER_UNAUTHORIZED_ERROR);
+    }
     const carRemoved = await CarsService.remove(Number(id));
     if (carRemoved.err) return next(carRemoved.err);
     return res.status(OK_STATUS).json(carRemoved);
@@ -42,8 +49,12 @@ class CarsController {
   async update(
     req: Request, res: Response, next: NextFunction,
   ): Promise<Response | void> {
-    const { carModel, costHour }: BaseCar = req.body;
     const { id } = req.params;
+    const { userRole, userId } = req.user;
+    if (userRole !== 'admin' && id !== userId) {
+      return res.status(UNAUTHORIZED_STATUS).json(USER_UNAUTHORIZED_ERROR);
+    }
+    const { carModel, costHour }: BaseCar = req.body;
     const carUpdated = await CarsService.update({
       carId: Number(id), carModel, costHour,
     });
